bot/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
twitchAPI==3.11.0
typing_extensions==4.6.3
yarl==1.9.2
